att/gatt: handle WriteCommand(Write without Response) properly
Currently, in the GATT API, we handle both case with the same
handler:
	type WriteHandler interface {
		ServeWrite(req Request, rsp ResponseWriter)
	}

Currently, when the handler was invoked to handle WriteCommand,
which doesn't have any use and hence passes nil as the rsp, user
may run into problem if he manipulates the rsp without checking
the request type.

Possible solutions:
a. Enforce user to check the request type.
b. Separate the API/Handler for Write and WriteCommand.
c. Create a dummy ResponseWriter for WriteCommand, which doesn't
   have effect in most operation.

We're taking the c) for now until we have better solution.