osx: unexport darwin.Advertisement
The bt.Advertisement interface has covered all the info
that corebluetooth exposes to us.

We keep hci.Advertisement exported because it not only implements
bt.Advertisement, but also contains lower-level HCI event that
is not exported by the interface. So user interested in those
info can use type assertion on bt.Advertisement to get it.