design: remove Server and Device
These are not used in the OS X implementation.
Not much meanning to keep them in the ble package, which supposed
to be common or shared between platforms.

examples/lib/gatt packages now can move on using linux/hci and
linux/gatt packages directly. This frees us more freedom to implement
linux-specific functions.