hci: fix Advertising Data and Scan Response handling.
AD and SR,if any, are delivered separately via HCI.
Upon recieving an AD, we'll create an Advertisement and pass it
to user's AdvHandler. If it's a scannable AD, which means there
will be a Scan Response coming, we cache this AD, so when the
SR is recieved, we associate it to the cached Advertisement,
and pass it to the user.

Normally, user will recieves Advertisement twice, the first one
has AD only, and the second one has AD and SR.