socket: unblock pending Read(), if any.
Any pending reference to the socket prevents the underlying HCI device
from being released.  This most likely happens due to a blocked Read().
The Close() syscall does not unblock the Read(), and the HCI socket does
not support Shutdown() syscall, which can be used for a TCP socket.
Here we just send a HCI Reset Command, which results a HCI response that
subsequently unblocks the Read().